// frontend/src/services/api/types.gen.ts
export type OrderData = {
    name: string;
    email: string;
    phoneNo: string;
    address: string;
    products: Array<{
        id?: number;
        quantity?: number;
    }>;
};